Rename vizualization.py to visualization.py and clean duplicate imports
Fixes #51
