+++ safe_ice/analysis/__init__.py
@@ -2,6 +2,6 @@
 """Analysis and visualization tools for Safe-ICE."""
 
 from .performance import PerformanceEvaluator
-from .vizualization import AdvancedAnalysis
+from .visualization import AdvancedAnalysis
 
 __all__ = ["PerformanceEvaluator", "AdvancedAnalysis"]
